requirements.txt
pyserial==3.5
pyspectator==1.2.2
